Commit 7 - Store and Tutorial pages
Page allow a user to buy an item with their coins, which runs the buy.php script. Script will check the user has enough coins/the item will have an effect before applying the changes to the character selected, and the total coins of the user. User who has no characters will now be redirected to a tutorials page upon logging in.